Catch horizontalDragUpdate in MaterialSeekBar
`PageView`s and `ListView`s (And most probably many other scrollable widgets) use the `Scrollable` widget which uses `HorizontalDragGestureRecognizer`. If any of the two widgets are used as a parent to `MaterialSeekBar`, horizontal drag gestures are caught by `GestureDetectors` used in these widgets containing a `Scrollable` since we only catch `Pan` gestures in `MaterialSeekBar`.

If one video is used for each page in a `PageView`, seeking the `MaterialSeekBar` leads to swiping and going to the next page because of this.